fix: fix dash while sliding by checking if on floor and adding velocity 🐛
If controlled node is not on floor, body will be set to FreeFallState
with no dash. To implement no dash, now exiting DashChargingState does
not set absolute velocity but adds to existing velocity. This also makes
dash logic better, for example dash will be slower if in opposite
direction to the movement of an already sliding body. This prevents a
loophole that allows a player to fly infinitely if they slide off a
cliff while charging their dash.