feat: add title screen and menus ✨
Following UI Screens are added:
	- Title Screen (with buttons to quit, go to developer website
	  and game website)
	- Play Screen
	- Host and Join Screens
		- Config options have not been implemented in host and
		  join screen and placeholder labels have been put
		  instead

*Additional Changes as a sideffect*
For adding menus, the MultiplayerManager was moved from a local scene to
being a direct child of the root node. This broke
MultiplayerSynchronizer, to fix which MultiplayerSpawner had to be used.
Player spawning logic also needed to be moved to MultiplayerSpawner's
spawn_function from MultiplayerManager.