feat: add health amd dash charge bar ✨
Health bar is shown always, and automatically updates when changing the
value of health using a setter function. This is also synced across
connected peers.

Dash charge bar shows how much dash has been charged, with an upper
limit. The dash charging bar is visible only to the player charging his
dash (not to other multiplayer peers) and only when they are in the dash
charging state